requirements.txt
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pandas==2.3.3
python-dotenv==1.1.1
jupyter==1.1.1
ipykernel==7.1.0
matplotlib==3.10.7
seaborn==0.13.2
numpy==2.3.0